dsp-processor/requirements.txt
# DSP Processor Dependencies
numpy>=1.21.0
scipy>=1.7.0
pyserial>=3.5
matplotlib>=3.5.0

# Desarrollo y testing
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
pylint>=2.17.0
